dev_requirements.txt
django-debug-toolbar
django-extensions
ipython[notebook]